NXP-25626: Upgrade CQ, reduce cq4 file size
Log metadata are persisted into a metadata.properties file,
the meta includes: the number of partitions, the retention and cycle to
use and the block size which determines the initial cq4 file size and
the maximum message. Default is 4MB block size to get 1MB records.
The codec is not included as metdata because it is not always known when
initializing the Log.